[Refactor: 좋아요 기능 구조 리팩토링] 좋아요 커스텀 훅 파일 삭제 후 api만 분리 형식으로 다시 수정
