+++ src/components/shared/likes/index.tsx
@@ -1,6 +1,7 @@
+import { useCallback, useState } from "react";
 import Icon from "../Icon"
 import { ActiveBox, Container } from "./styles"
-import { useLike } from "@/hooks/useLike";
+import { updateLike } from "@/apis";
 
 interface LikesProps {
     postId: number;
@@ -8,16 +9,39 @@ interface LikesProps {
     likeType : boolean;
 }
 
-const Likes: React.FC<LikesProps> = ({ postId, likeCount: initialLikeCount, likeType: initialLikeType }) => {
+const Likes: React.FC<LikesProps> = ({ postId, likeCount, likeType }) => {
+  const [nLikeCount, setNLikeCount] = useState(likeCount);
+  const [nLikeType, setNLikeType] = useState(likeType);
+  const [jellyAnimation, setJellyAnimation] = useState(false);
 
-  const { likeCount, likeType, jellyAnimation, toggleLike } = useLike( postId, initialLikeCount, initialLikeType );
+  const toggleLike = useCallback(async () => {
+    // 선 렌더링 (animation, 좋아요 변화 반영)
+    const newLikeType = !nLikeType;
+    setNLikeType(newLikeType);
+    setNLikeCount((prev) => (newLikeType ? prev + 1 : prev - 1));
+    setJellyAnimation((prev) => !prev);
+
+    setTimeout(() => {
+      setJellyAnimation((prev) => !prev);
+    }, 100);
+
+    // 상태 업데이트
+    try {
+      await updateLike(postId);
+    } catch (error) {
+      console.log('실패', error);
+      setNLikeType(nLikeType);
+      setNLikeCount((prev) => (newLikeType ? prev - 1 : prev + 1));
+    }
+    // likeType과 postId값이 변하지 않는 이상 함수는 재생성 x
+  }, [nLikeType, postId]);
 
   return (
     <Container>
       <ActiveBox onClick={toggleLike} className={jellyAnimation ? '' : 'jelly'}>
-        <Icon value={likeType ? "yeslike" : "nolike"} />
+        <Icon value={nLikeType ? "yeslike" : "nolike"} />
       </ActiveBox>
-      <p>{likeCount}</p>
+      <p>{nLikeCount}</p>
     </Container>
   )
 }
